requirements.txt
altgraph==0.17
future==0.18.2
pefile==2019.4.18
psycopg2-binary==2.8.6
pyinstaller==4.1
pyinstaller-hooks-contrib==2020.11
pywin32-ctypes==0.2.0
